+++ app/adms/Views/Users/viewUser.php
@@ -11,7 +11,7 @@ if (!empty($this->data['viewUser'])) {
     echo "<a href=' " . URLADM . "edit-users/index/" . $this->data['viewUser'][0]['id'] . "'>Editar</a><br>";
     echo "<a href=' " . URLADM . "edit-users-password/index/" . $this->data['viewUser'][0]['id'] . "'>Editar Senha</a><br>";
     echo "<a href=' " . URLADM . "edit-users-image/index/" . $this->data['viewUser'][0]['id'] . "'>Editar Imagem</a><br>";
-    echo "<a href=' " . URLADM . "delete-users/index/" . $this->data['viewUser'][0]['id'] . "'>Excluir</a><br><br>";
+    echo "<a href=' " . URLADM . "delete-users/index/" . $this->data['viewUser'][0]['id'] . "' onClick='return confirm(\"Tem Certeza que deseja excluir este registro?\")'>Excluir</a><br><br>";
 }
 
 if (isset($_SESSION['msg'])) {
